Apply the same `box' thing as before, to allow frame mutation.
Note that this is not yet complete: there's also a need to mutate
*bindings*, not just the frames.  This is the leftover `set-car!'  that
is still in the code.  It also shows why there needs to be some
clarification on what is a "frame": the current version of the code
creates new frames while scanning bindings, and when the box is there,
this is a bad hack, since you wouldn't want to allow mutating these
temporary frames.

Note that I also changed the cosmetic look of `add-binding!', because it
looks like this code is a copy of
"env-functional-interpreter-compiler/env-ds.rkt", and so it seems like
it's good to minimize the differences between them.