Adapt existing tests to use `test'.
This is more verbose than the previous version, but that's because
`test' is more general.  Specifically, it throws an error with the
actual test *syntax* that was used -- for example,

  (test (+ 1 2) => 4)

works as expected, but if we do this with a testing utility:

  (define (test-add x y z)
    (test (+ x y) => z))
  (test-add 1 2 4)

then the error will show the actual expression:

  (+ x y): expected 4, got 3

and we don't see the actual values that were used.